+++ index.html
@@ -124,7 +124,7 @@
           <label class="ad-form__label" for="room_number">Количество комнат</label>
           <select id="room_number" name="rooms">
             <option value="1" selected>1 комната</option>
-            <option value="2" d>2 комнаты</option>
+            <option value="2">2 комнаты</option>
             <option value="3">3 комнаты</option>
             <option value="100">100 комнат</option>
           </select>
