+++ js/form.js
@@ -1,49 +1,51 @@
-import {addClassName, disableField, enableField, removeClassName} from './util.js';
-import {filter, filterFieldsets} from './filter.js';
-
 const form = document.querySelector('.ad-form');
-const fieldsets = form.querySelectorAll('fieldset');
+const formFieldsets = form.querySelectorAll('fieldset');
 const capacityField = form.querySelector('#capacity');
 const capacityOptions = capacityField.querySelectorAll('option');
 
-// Валидация формы
 
 function setAvailableCapacity () {
   const roomsNumber = Number(form.querySelector('#room_number').value);
-  const oneGuest = capacityOptions[2];
-  const twoGuests = capacityOptions[1];
-  const threeGuests = capacityOptions[0];
-  const zeroGuests = capacityOptions[3];
   const selectedGuestNumber = Number(capacityField.value);
 
   capacityOptions.forEach( (element) => {
     element.disabled = true;
+    const optionValue = Number(element.value);
+
+    switch (roomsNumber) {
+      case 1:
+        if (optionValue === 1) {
+          element.disabled = false;
+          element.selected = true;
+        }
+        break;
+      case 2:
+        if (optionValue === 1 || optionValue === 2) {
+          element.disabled = false;
+          if (selectedGuestNumber === 3 || selectedGuestNumber === 0) {
+            if (optionValue === 2) {
+              element.selected = true;
+            }
+          }
+        }
+        break;
+      case 3:
+        if (optionValue === 1 || optionValue === 2 || optionValue === 3) {
+          element.disabled = false;
+        }
+        if (selectedGuestNumber === 0) {
+          if (optionValue === 3) {
+            element.selected = true;
+          }
+        }
+        break;
+      default:
+        if (optionValue === 0) {
+          element.disabled = false;
+          element.selected = true;
+        }
+    }
   });
-
-  if (roomsNumber === 1) {
-    oneGuest.disabled = false;
-    oneGuest.selected = true;
-  } else if (roomsNumber === 2) {
-    oneGuest.disabled = false;
-    twoGuests.disabled = false;
-
-    if (selectedGuestNumber === 3 || selectedGuestNumber === 0) {
-      twoGuests.selected = true;
-    }
-
-  } else if (roomsNumber === 3) {
-    oneGuest.disabled = false;
-    twoGuests.disabled = false;
-    threeGuests.disabled = false;
-
-    if (selectedGuestNumber === 0) {
-      threeGuests.selected = true;
-    }
-
-  } else {
-    zeroGuests.disabled = false;
-    zeroGuests.selected = true;
-  }
 }
 
 
@@ -57,33 +59,6 @@ function onAdFormChange (evt) {
 }
 
 
-// Активация и деактивация формы
-
-function deactivateForms () {
-  form.removeEventListener('change', onAdFormChange);
-  addClassName(form, 'ad-form--disabled');
-  addClassName(filter, 'map__filters--disabled');
-  fieldsets.forEach( (element) => {
-    disableField(element);
-  } );
-  filterFieldsets.forEach( (element) => {
-    disableField(element);
-  } );
-}
-
-function activateForms () {
-  form.addEventListener('change', onAdFormChange);
-  setAvailableCapacity();
-  removeClassName(form, 'ad-form--disabled');
-  removeClassName(filter, 'map__filters--disabled');
-  fieldsets.forEach( (element) => {
-    enableField(element);
-  } );
-  filterFieldsets.forEach( (element) => {
-    enableField(element);
-  } );
-}
-
-export {deactivateForms, activateForms};
+export {form, formFieldsets, setAvailableCapacity, onAdFormChange};
